VLBISkyRegularizers

Documentation for VLBISkyRegularizers.

VLBISkyRegularizers.ALRDomain

VLBISkyRegularizers.AbstractRegularizer

VLBISkyRegularizers.AddRegularizer

VLBISkyRegularizers.CLRDomain

VLBISkyRegularizers.L1

VLBISkyRegularizers.LinearDomain

VLBISkyRegularizers.LogDomain

VLBISkyRegularizers.ParameterDomain

VLBISkyRegularizers.TSV

VLBISkyRegularizers.TV

VLBISkyRegularizers.WaveletL1

VLBISkyRegularizers.WeightRegularizer

Distributions._logpdf

VLBISkyRegularizers._transform_linear

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.evaluation_domain

VLBISkyRegularizers.grid

VLBISkyRegularizers.image_domain

VLBISkyRegularizers.l1_base

VLBISkyRegularizers.l1_base

VLBISkyRegularizers.regularizers

VLBISkyRegularizers.transform_domain

VLBISkyRegularizers.transform_image

VLBISkyRegularizers.tsv_base

VLBISkyRegularizers.tsv_base

VLBISkyRegularizers.tv_base

VLBISkyRegularizers.tv_base

VLBISkyRegularizers.wavelet_transform

VLBISkyRegularizers.ALRDomain — Type

ALRDomain

Additive Log-Ratio transform. Can only be used as an image domain. Image transformation is the inverse ALR function. Similar to softmax but treats one pixel (the last pixel) as special.

VLBISkyRegularizers.AbstractRegularizer — Type

AbstractRegularizer

Used to define the regularizer functions. See subtypes(AbstractRegularizer) for a list of implemented regularizers.

Mandatory fields

hyperparameter::Number: the hyperparameter of the regularization function.

image_domain::AbstractDomain: the domain of the image space

evaluation_domain::AbstractDomain: the domain on which the regularizer is to be evaluated

grid: grid on which image is defined

VLBISkyRegularizers.AddRegularizer — Type

AddRegularizer <: AbstractRegularizer

Structure for adding two regularizers. The two regularizers must have the same grid and image domain.

VLBISkyRegularizers.CLRDomain — Type

CLRDomain

Centered Log-Ratio transformation. Can only be used as an image domain. Image transformation is the softmax function (applies the exponential function and normalizes by the sum).

VLBISkyRegularizers.L1 — Type

L1 <: AbstractRegularizer

Regularizer using the L1 norm.

fields

hyperparameter::Number: the hyperparameter of the regularization function.

image_domain::AbstractDomain: the domain of the image space

evaluation_domain::AbstractDomain: the domain on which the regularizer is to be evaluated

grid: grid on which image is defined

VLBISkyRegularizers.LinearDomain — Type

LinearDomain

Can be used as an image and evaluation domain. No transformation in evaluation. Image transformation normalizes the image by the sum and absolute value.

VLBISkyRegularizers.LogDomain — Type

LogDomain

Can be used as an image and evaluation domain. Logarithmic transformation in evaluation. Image transformation is the softmax function (applies the exponential function and normalizes by the sum).

VLBISkyRegularizers.ParameterDomain — Type

ParameterDomain

Can only be used as an evaluation domain. Sets the evaluation domain equal to the image domain.

VLBISkyRegularizers.TSV — Type

TSV <: AbstractRegularizer

Regularizer using the Isotropic Total Squared Variation.

fields

hyperparameter::Number: the hyperparameter of the regularization function.

image_domain::AbstractDomain: the domain of the image space

evaluation_domain::AbstractDomain: the domain on which the regularizer is to be evaluated

grid: grid on which image is defined

VLBISkyRegularizers.TV — Type

TV <: AbstractRegularizer

Regularizer using the Isotropic Total Variation.

fields

hyperparameter::Number: the hyperparameter of the regularization function.

image_domain::AbstractDomain: the domain of the image space

evaluation_domain::AbstractDomain: the domain on which the regularizer is to be evaluated

grid: grid on which image is defined

VLBISkyRegularizers.WaveletL1 — Type

WaveletL1 <: AbstractRegularizer

Regularizer using the l1-norm with a wavelet transform

fields

hyperparameter::Number: the hyperparameter of the regularization function.

image_domain::AbstractDomain: the domain of the image space

evaluation_domain::AbstractDomain: the domain on which the regularizer is to be evaluated

wavelet::WVType: wavelet type

grid: grid on which image is defined

VLBISkyRegularizers.WeightRegularizer — Type

WeightRegularizer <: AbstractRegularizer

Structure for weighting a regularizer by a scalar.

Distributions._logpdf — Method

Distributions._logpdf(d::Regularizers, image::AbstractMatrix{<:Real})

The log density of the regularizers evaluated at the input image. .

Arguments

reg::Regularizers: the regularizer functions.

image::AbstractMatrix{<:Real}: the model of the input image

VLBISkyRegularizers._transform_linear — Method

_transform_linear(::AbstractDomain, x::AbstractArray)

Transforms an array to the linear domain. Same as transform_image but does not necessarily impose non-negativity and normalization.

VLBISkyRegularizers.evaluate — Method

evaluate(::AbstractRegularizer, ::AbstractArray)

Compute the value of the given regularization function on the input image. By default, return 0.

VLBISkyRegularizers.evaluate — Method

evaluate(reg::L1, x::AbstractArray)

Evaluate the L1 norm regularizer at an image.

Arguments

reg::L1: L1 norm regularizer

x::AbstractArray: the image

VLBISkyRegularizers.evaluate — Method

evaluate(reg::TSV, x::AbstractArray)

Evaluate the TSV regularizer at an image.

Arguments

reg::TSV: TSV regularizer

x::AbstractArray: the image

VLBISkyRegularizers.evaluate — Method

evaluate(reg::TV, x::AbstractArray)

Evaluate the TV regularizer at an image.

Arguments

reg::TV: TV regularizer

x::AbstractArray: the image

VLBISkyRegularizers.evaluation_domain — Method

evaluation_domain(reg::AbstractRegularizer)

Return the evaluation domain of the given regularizer.

VLBISkyRegularizers.grid — Method

grid(reg::AbstractRegularizer)

Return the grid on which an image to be regularized is defined.

VLBISkyRegularizers.image_domain — Method

image_domain(reg::AbstractRegularizer)

Return the image domain of the given regularizer.

VLBISkyRegularizers.l1_base — Method

l1_base(x::AbstractArray, w::Number)

Base function of the L1 norm.

Arguments

x::AbstractArray: the image

'w::Number' : the regularization weight

VLBISkyRegularizers.l1_base — Method

l1_base(x::AbstractArray)

Base function of the L1 norm.

Arguments

x::AbstractArray: the image

VLBISkyRegularizers.regularizers — Method

regularizers(::AbstractRegularizer)

List regularizers used.

VLBISkyRegularizers.transform_domain — Method

transform_domain(::LinearDomain, x::AbstractArray)

Transform the domain of the image to the evaluation domain.

VLBISkyRegularizers.transform_image — Method

transform_image(::AbstractDomain, x::AbstractArray)

Transforms an array to the normalized, non-negative image domain.

VLBISkyRegularizers.tsv_base — Method

tsv_base(x::AbstractArray, w::Number)

Base function of the isotropic total squared variation.

Arguments

x::AbstractArray: the image

'w::Number' : the regularization weight

VLBISkyRegularizers.tsv_base — Method

tsv_base(x::AbstractArray)

Base function of the isotropic total squared variation.

Arguments

x::AbstractArray: the image

VLBISkyRegularizers.tv_base — Method

tv_base(x::AbstractArray, w::Number)

Base function of the isotropic total variation.

Arguments

x::AbstractArray: the image

'w::Number' : the regularization weight

VLBISkyRegularizers.tv_base — Method

tv_base(x::AbstractArray)

Base function of the isotropic total variation.

Arguments

x::AbstractArray: the image

VLBISkyRegularizers.wavelet_transform — Method

wavelet_l1_base(x::AbstractArray, wv::WaveletL1)

Base function of the Wavelet L1 norm.

Arguments

x::AbstractArray: the image

wv::WVType : the wavelet transform type