VLBISkyRegularizers

Documentation for VLBISkyRegularizers.

VLBISkyRegularizers.AbstractRegularizer

VLBISkyRegularizers.KLEntropy

VLBISkyRegularizers.L1Norm

VLBISkyRegularizers.TSV

VLBISkyRegularizers.TV

VLBISkyRegularizers.domain

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.hyperparameter

VLBISkyRegularizers.klentropy_base

VLBISkyRegularizers.l1norm

VLBISkyRegularizers.l1norm

VLBISkyRegularizers.tsv_base

VLBISkyRegularizers.tsv_base

VLBISkyRegularizers.tsv_base_pixel

VLBISkyRegularizers.tv_base

VLBISkyRegularizers.tv_base

VLBISkyRegularizers.tv_base_pixel

VLBISkyRegularizers.AbstractRegularizer — Type

AbstractRegularizer

Mandatory fields

hyperparameter::Number: the hyper parameter of the regularization function. In default, it should be a number.

domain::RegularizerDomain: the domain of the image space where the regularization function will be computed.

Mandatory methods

evaluate(::AbstractRegularizerDomain, ::AbstractRegularizer, ::AbstractImageModel, ::AbstractArray): Evaluate the regularization function for the given input sky image.

evaluate(::AbstractRegularizer, ::AbstractImageModel, ::AbstractArray): Evaluate the regularization function for the given input sky image.

cost(::AbstractRegularizer, ::AbstractImageModel, ::AbstractArray): Evaluate the cost function for the given input sky image. The cost function is defined by the product of its hyperparameter and regularization function.

``

VLBISkyRegularizers.KLEntropy — Type

KLEntropy <: AbstractRegularizer

Regularizer using the KL divergence.

fields

hyperparameter::Number: the hyperparameter of the regularizer

prior: the prior image

domain::AbstractRegularizerDomain: the image domain where the regularization funciton will be computed. L1Norm can be computed only in LinearDomain().

VLBISkyRegularizers.L1Norm — Type

L1Norm <: AbstractRegularizer

Regularizer using the l1-norm.

fields

hyperparameter::Number: the hyperparameter of the regularizer

weight: the weight of the regularizer, which could be a number or an array.

domain::AbstractRegularizerDomain: the image domain where the regularization funciton will be computed. L1Norm can be computed only in LinearDomain().

VLBISkyRegularizers.TSV — Type

TSV <: AbstractRegularizer

Regularizer using the Isotropic Total Squared Variation.

fields

hyperparameter::Number: the hyperparameter of the regularizer

weight: the weight of the regularizer, which could be a number or an array.

domain::AbstractRegularizerDomain: the image domain where the regularization funciton will be computed. L1Norm can be computed only in LinearDomain().

VLBISkyRegularizers.TV — Type

TV <: AbstractRegularizer

Regularizer using the Isotropic Total Variation.

fields

hyperparameter::Number: the hyperparameter of the regularizer

weight: the weight of the regularizer, which could be a number or an array.

domain::AbstractRegularizerDomain: the image domain where the regularization funciton will be computed. L1Norm can be computed only in LinearDomain().

VLBISkyRegularizers.domain — Method

domain(reg::AbstractRegularizer) = reg.domain

Return the computing domain of the given regularizer.

VLBISkyRegularizers.evaluate — Method

evaluate(::AbstractRegularizer, skymodel::AbstractImage2DModel, x::AbstractArray)

VLBISkyRegularizers.evaluate — Method

evaluate(::AbstractRegularizer, skymodel::AbstractImage2DModel, x::AbstractArray)

VLBISkyRegularizers.hyperparameter — Method

hyperparameter(reg::AbstractRegularizer) = reg.hyperparameter

Return the hyperparameter of the given regularizer.

VLBISkyRegularizers.klentropy_base — Method

klentropy_base(I::IntensityMap, p::IntensityMap)

Base function of the l1norm.

Arguments

I::IntensityMap: the image

p::IntensityMap: the prior image

VLBISkyRegularizers.l1norm — Method

l1norm(I::IntensityMap, w::Number)

Base function of the l1norm.

Arguments

I::IntensityMap: the image

w::Number: the regularization weight

VLBISkyRegularizers.l1norm — Method

l1norm(I::IntensityMap)

Base function of the l1norm.

Arguments

I::IntensityMap: the image

VLBISkyRegularizers.tsv_base — Method

tsv_base(I::IntensityMap, w::Number)

Base function of the isotropic total squared variation.

Arguments

I::IntensityMap: the image

'w::Number' : the regularization weight

VLBISkyRegularizers.tsv_base — Method

tsv_base(I::IntensityMap)

Base function of the isotropic total squared variation.

Arguments

I::IntensityMap: the image

VLBISkyRegularizers.tsv_base_pixel — Method

tsv_base_pixel(I::AbstractArray, ix::Integer, iy::Integer)

Evaluate the isotropic squared variation term for the given pixel

Arguments

I::IntensityMap: the image

VLBISkyRegularizers.tv_base — Method

tv_base(I::IntensityMap, w::Number)

Base function of the isotropic total variation.

Arguments

I::IntensityMap: the image

'w::Number' : the regularization weight

VLBISkyRegularizers.tv_base — Method

tv_base(I::IntensityMap)

Base function of the isotropic total variation.

Arguments

I::IntensityMap: the image

VLBISkyRegularizers.tv_base_pixel — Method

tv_base_pixel(I::AbstractArray, ix::Integer, iy::Integer)

Evaluate the isotropic variation term for the given pixel

Arguments

I::IntensityMap: the image