VLBISkyRegularizers

Documentation for VLBISkyRegularizers.

VLBISkyRegularizers.AbstractRegularizer

VLBISkyRegularizers.L1

VLBISkyRegularizers.Regularizers

VLBISkyRegularizers.TSV

VLBISkyRegularizers.TV

Distributions._logpdf

Distributions._rand!

VLBISkyRegularizers.domain

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.evaluate

VLBISkyRegularizers.hyperparameter

VLBISkyRegularizers.l1_base

VLBISkyRegularizers.l1_base

VLBISkyRegularizers.transform_domain

VLBISkyRegularizers.transform_domain

VLBISkyRegularizers.tsv_base

VLBISkyRegularizers.tsv_base

VLBISkyRegularizers.tv_base

VLBISkyRegularizers.tv_base

VLBISkyRegularizers.AbstractRegularizer — Type

AbstractRegularizer

Mandatory fields

hyperparameter::Number: the hyper parameter of the regularization function. In default, it should be a number.

domain::AbstractRegularizerDomain: the domain of the image space where the regularization function will be computed.

Mandatory methods

evaluate(::RegularizerDomain, ::Regularizer, ::ImageModel, ::AbstractArray): Evaluate the regularization function for the given input sky image.

evaluate(::Regularizer, ::ImageModel, ::AbstractArray): Evaluate the regularization function for the given input sky image.

cost(::AbstractRegularizer, ::AbstractImageModel, ::AbstractArray): Evaluate the cost function for the given input sky image. The cost function is defined by the product of its hyperparameter and regularization function.

``

VLBISkyRegularizers.L1 — Type

L1 <: AbstractRegularizer

Regularizer using the L1 norm.

fields

hyperparameter::Number: the hyperparameter of the regularizer

domain::AbstractRegularizerDomain: the image domain where the regularization funciton will be computed.

VLBISkyRegularizers.Regularizers — Type

Regularizers

Mandatory fields

regularizers::NTuple{N, AbstractRegularizer}: Tuple of AbstractRegularizer objects

grid: grid on which the image is defined

VLBISkyRegularizers.TSV — Type

TSV <: AbstractRegularizer

Regularizer using the Isotropic Total Squared Variation.

fields

hyperparameter::Number: the hyperparameter of the regularizer

domain::AbstractRegularizerDomain: the image domain where the regularization funciton will be computed.

VLBISkyRegularizers.TV — Type

TV <: AbstractRegularizer

Regularizer using the Isotropic Total Variation.

fields

hyperparameter::Number: the hyperparameter of the regularizer

domain::AbstractRegularizerDomain: the image domain where the regularization funciton will be computed.

Distributions._logpdf — Method

Distributions._logpdf(d::Regularizers, image::AbstractMatrix{<:Real})

The log density of the regularizers evaluated at the input image. .

Arguments

reg::Regularizers: the regularizer functions.

image::AbstractMatrix{<:Real}: the model of the input image

Distributions._rand! — Method

Distributions._rand!(rng::Random.AbstractRNG, ::Regularizers, x::AbstractMatrix)

Return a random sample of shape equal to the shape of the input matrix

Arguments

rng::Random.AbstractRNG: an RNG seed object

::Regularizers: any set of regularizer functions

x::AbstractMatrix: input matrix

VLBISkyRegularizers.domain — Method

domain(reg::AbstractRegularizer) = reg.domain

Return the computing domain of the given regularizer.

VLBISkyRegularizers.evaluate — Method

evaluate(reg::L1, x::AbstractArray)

Evaluate the L1 norm regularizer at an image.

Arguments

reg::L1: L1 norm regularizer

x::AbstractArray: the image

VLBISkyRegularizers.evaluate — Method

evaluate(reg::TSV, x::AbstractArray)

Evaluate the TSV regularizer at an image.

Arguments

reg::TSV: TSV regularizer

x::AbstractArray: the image

VLBISkyRegularizers.evaluate — Method

evaluate(reg::TV, x::AbstractArray)

Evaluate the TV regularizer at an image.

Arguments

reg::TV: TV regularizer

x::AbstractArray: the image

VLBISkyRegularizers.evaluate — Method

evaluate(::AbstractRegularizer, ::AbstractArray)

Compute the value of the given regularization function on the input image. By default, return 0.

VLBISkyRegularizers.hyperparameter — Method

hyperparameter(reg::AbstractRegularizer) = reg.hyperparameter

Return the hyperparameter of the given regularizer.

VLBISkyRegularizers.l1_base — Method

l1_base(x::AbstractArray, w::Number)

Base function of the L1 norm.

Arguments

x::AbstractArray: the image

'w::Number' : the regularization weight

VLBISkyRegularizers.l1_base — Method

l1_base(x::AbstractArray)

Base function of the L1 norm.

Arguments

x::AbstractArray: the image

VLBISkyRegularizers.transform_domain — Method

transform_domain(::LinearDomain, x::AbstractArray)

Transform the domain of the image. For Linear Domain, this does not change the image.

VLBISkyRegularizers.transform_domain — Method

transform_domain(::LogDomain, x::AbstractArray)

Transform the domain of the image. For Log Domain, return the log of the image.

VLBISkyRegularizers.tsv_base — Method

tsv_base(x::AbstractArray, w::Number)

Base function of the isotropic total squared variation.

Arguments

x::AbstractArray: the image

'w::Number' : the regularization weight

VLBISkyRegularizers.tsv_base — Method

tsv_base(x::AbstractArray)

Base function of the isotropic total squared variation.

Arguments

x::AbstractArray: the image

VLBISkyRegularizers.tv_base — Method

tv_base(x::AbstractArray, w::Number)

Base function of the isotropic total variation.

Arguments

x::AbstractArray: the image

'w::Number' : the regularization weight

VLBISkyRegularizers.tv_base — Method

tv_base(x::AbstractArray)

Base function of the isotropic total variation.

Arguments

x::AbstractArray: the image